create package-card and move files
