+++ app/src/main/java/ksnd/open/hiraganaconverter/view/screen/ConverterScreen.kt
@@ -52,7 +52,7 @@ import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.google.accompanist.systemuicontroller.rememberSystemUiController
 import ksnd.open.hiraganaconverter.R
-import ksnd.open.hiraganaconverter.view.parts.ConversionTypeSpinnerCard
+import ksnd.open.hiraganaconverter.view.parts.card.ConversionTypeSpinnerCard
 import ksnd.open.hiraganaconverter.view.parts.TopBar
 import ksnd.open.hiraganaconverter.view.parts.button.ConvertButton
 import ksnd.open.hiraganaconverter.view.parts.button.CustomFilledTonalIconButton
